+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Implement edit mapping profile feature. UIDEXP-132.
 * Provide translations for the mapping profile transformation field names. UIDEXP-140.
 * Update endpoints names according to changes on backend side. UIDEXP-146.
+* Add validation between selected record types and transformations on form submit. UIDEXP-138.
 
 ## [2.0.1](https://github.com/folio-org/ui-data-export/tree/v2.0.1) (2020-07-09)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v2.0.0...v2.0.1)